backend/article_creator/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
boto3==1.35.34
botocore==1.35.34
certifi==2024.8.30
distro==1.9.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
jmespath==1.0.1
numpy==2.1.1
openai==1.51.0
opencv-python==4.10.0.84
pillow==10.4.0
pip==24.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
